added lecture7 hometask fixes
